Fixed int/unsigned int warning in timing_place.c
git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@4013 8e3573b8-cf2c-4f14-ef6d-137439e28b8b

+++ vpr/SRC/place/timing_place.c
@@ -91,7 +91,7 @@ void load_criticalities(t_slack * slacks, float crit_exponent) {
 #endif
 
 	for (inet = 0; inet < g_clbs_nlist.net.size(); inet++) {
-		if (inet == OPEN)
+		if ((int) inet == OPEN)
 			continue;
 		if (g_clbs_nlist.net[inet].is_global)
 			continue;
